several changes applied to UI after test and review
