+++ app/src/main/res/layout/fragment_payment_with_token_layout.xml
@@ -21,8 +21,6 @@
         android:id="@+id/buyButton"
         style="@style/SdkDemoButtonStyle"
         android:layout_marginTop="@dimen/space_x2"
-        android:layout_height="0dp"
-        android:layout_width="match_parent"
         android:text="@string/button_continue_with_token_text" />
 
 </LinearLayout>
