+++ app/src/main/res/values/styles.xml
@@ -6,6 +6,7 @@
         <item name="colorPrimaryDark">@color/colorPrimaryDark</item>
         <item name="colorAccent">@color/colorAccent</item>
         <item name="toolbarStyle">@style/SdkDemoToolBarStyle</item>
+        <item name="android:windowLightStatusBar">true</item>
     </style>
 
     <style name="MobileSdkStyle" parent="Theme.MobileSdk.Light"></style>
@@ -32,7 +33,8 @@
 
     <style name="SdkDemoButtonStyle" parent="Widget.MaterialComponents.Button">
         <item name="fontFamily">@font/inter_regular</item>
-        <item name="android:minHeight">44dp</item>
+        <item name="android:layout_height">64dp</item>
+        <item name="android:layout_width">match_parent</item>
         <item name="cornerRadius">4dp</item>
         <item name="android:textStyle">bold</item>
         <item name="android:textColor">@drawable/button_text_selector</item>
